requirements.txt
opencv-python
customtkinter
Pillow
imageio
CTkMessagebox